Perf: 변경 케이크, 게시글 좋아요 exception 추가

+++ src/main/java/com/project/makecake/service/CakeService.java
@@ -133,8 +133,17 @@ public class CakeService {
         Cake foundCake = cakeRepository.findById(cakeId)
                 .orElseThrow(()->new IllegalArgumentException("케이크가 존재하지 않습니다."));
 
+        // 케이크 좋아요 찾기
+        Optional<CakeLike> foundCakeLike = cakeLikeRepository.findByUserAndCake(user,foundCake);
+
         // myLike가 true이면 새로운 cakeLike 저장
         if (requestDto.isMyLike()) {
+
+            // 이미 좋아요를 누른 케이크이면 exception
+            if (foundCakeLike.isPresent()) {
+                throw new IllegalArgumentException("이미 좋아요를 누른 케이크입니다.");
+            }
+
             CakeLike cakeLike = CakeLike.builder()
                     .cake(foundCake)
                     .user(user)
@@ -143,6 +152,12 @@ public class CakeService {
 
         // myLike가 false이면 기존 cakeLike 삭제
         } else {
+
+            // 좋아요를 누르지 않은 케이크이면 exception
+            if (!foundCakeLike.isPresent()) {
+                throw new IllegalArgumentException("좋아요를 누르지 않은 케이크입니다.");
+            }
+
             cakeLikeRepository.deleteByUserAndCake(user, foundCake);
         }
